requirements.txt
fastapi
uvicorn[standard]
sqlmodel
playwright
google-generativeai
apscheduler
python-dotenv
pydantic-settings
beautifulsoup4
